requirements.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
pillow
runpod
